refactor: application 패키지 애그리거트 분할 및 이동 #12
