+++ src/main/java/tobyspring/splearn/adapter/integration/DummyEmailSender.java
@@ -2,7 +2,7 @@ package tobyspring.splearn.adapter.integration;
 
 import org.springframework.context.annotation.Fallback;
 import org.springframework.stereotype.Component;
-import tobyspring.splearn.application.required.EmailSender;
+import tobyspring.splearn.application.member.required.EmailSender;
 import tobyspring.splearn.domain.shared.Email;
 
 @Component
