+++ src/main/java/tobyspring/splearn/application/member/provided/MemberFinder.java
@@ -1,4 +1,4 @@
-package tobyspring.splearn.application.provided;
+package tobyspring.splearn.application.member.provided;
 
 import tobyspring.splearn.domain.member.Member;
