+++ src/main/java/tobyspring/splearn/application/member/provided/MemberRegister.java
@@ -1,4 +1,4 @@
-package tobyspring.splearn.application.provided;
+package tobyspring.splearn.application.member.provided;
 
 import jakarta.validation.Valid;
 import tobyspring.splearn.domain.member.Member;
