+++ src/test/java/tobyspring/splearn/SplearnTestConfiguration.java
@@ -2,7 +2,7 @@ package tobyspring.splearn;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import tobyspring.splearn.application.required.EmailSender;
+import tobyspring.splearn.application.member.required.EmailSender;
 import tobyspring.splearn.domain.MemberFixture;
 import tobyspring.splearn.domain.member.PasswordEncoder;
